@@ -3,30 +3,31 @@
   "license": "EUPL-1.2",
   "type": "module",
   "scripts": {
-    "format": "prettier --write --plugin prettier-plugin-svelte .",
+    "format": "prettier --write .",
     "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview",
-    "check": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json",
-    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json --watch",
-    "clean": "rimraf build .svelte-kit"
+    "check": "svelte-kit sync && svelte-check --no-tsconfig",
+    "check:watch": "svelte-kit sync && svelte-check --no-tsconfig --watch"
   },
   "dependencies": {
     "@minvws/manon": "workspace:*",
-    "@minvws/manon-themes": "workspace:*"
+    "@minvws/manon-themes": "workspace:*",
+    "highlight.js": "^11.11.1"
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "^2.0.3",
-    "@sveltejs/kit": "^1.30.3",
-    "prettier": "^3.1.1",
-    "prettier-plugin-svelte": "^3.1.2",
-    "rimraf": "^5.0.5",
-    "sass": "^1.69.5",
-    "svelte": "^4.2.8",
-    "svelte-check": "^3.6.2",
-    "svelte-highlight": "^7.4.2",
+    "@sveltejs/adapter-static": "^3.0.9",
+    "@sveltejs/kit": "^2.43.5",
+    "@sveltejs/vite-plugin-svelte": "^6.2.1",
+    "mdsvex": "^0.12.6",
+    "prettier": "^3.6.2",
+    "prettier-plugin-svelte": "^3.4.0",
+    "sass": "^1.93.2",
+    "svelte": "^5.39.6",
+    "svelte-check": "^4.3.2",
     "svelte-preprocess-import-assets": "^1.1.0",
-    "typescript": "^5.3.3",
-    "vite": "^4.5.1"
+    "typescript": "^5.9.2",
+    "unist-util-visit": "^5.0.0",
+    "vite": "^7.1.7"
   }
 }
